+++ mods/autopilot/src/vad/types.ts
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-type Vad = (chunk: Float32Array, cb: (event: string) => void) => Promise<void>;
+type Vad = (chunk: Uint8Array, cb: (event: string) => void) => Promise<void>;
 
 type SpeechProbabilities = {
   notSpeech: number;
